product/requirements.txt
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2==2.7.4
django-extensions==3.1.3
web3==5.25.0
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.2
eth-event==1.2.3
eth-hash==0.3.2
